removed the functionality of the noSpamBot
This Bot has a fundamental problem:
It tries to send messages to users that have never interacted with the bot.
Therefore it will receive 403 errors.
Because of bad error handling the bot tries to resend the request as fast as possible in an endless loop.
This loop of 403 and 400 Errors occurs at about 10 Errors/s.
Because of this my internet traffic will rise to 6 Mbit/s and the sheer mass of requests overloads the processor of my modem and increases ping and package loss drastically.